requirements.txt
# AI女书应用 - 最小依赖版本 (保持所有功能)
# 核心Web框架
Flask==3.1.0
Werkzeug==3.1.3
Jinja2==3.1.5
MarkupSafe==2.1.5
itsdangerous==2.2.0
click==8.1.8
blinker==1.9.0
flask-cors==4.0.0

# AI和机器学习核心依赖
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
transformers==4.48.1
scikit-learn==1.6.1
numpy==1.26.4
scipy==1.13.1

# AI服务API
zhipuai==2.1.5.20250825

# 网络优化
httpx==0.25.2
urllib3==2.0.7
requests==2.31.0

# 图像和视频处理
Pillow==11.2.1
opencv-python==4.11.0.86
moviepy==1.0.3
imageio==2.37.0
imageio-ffmpeg==0.6.0

# 文本处理和翻译
googletrans-py==4.0.0
jieba==0.42.1
gensim==4.3.3

# 网络请求和配置
requests==2.32.3
httpx==0.28.1
urllib3==2.3.0
python-dotenv==1.0.1

# 文件处理和工具
filetype==1.2.0
ffmpy==0.5.0
uuid==1.30
python-dateutil==2.9.0.post0
pydub==0.25.1

# 数据序列化
PyYAML==6.0.2

# 其他必要依赖
certifi==2024.12.14
charset-normalizer==3.4.1
idna==2.10
packaging==24.2
six==1.17.0
tqdm==4.67.1